예약 도메인 수정

+++ storeReservation/src/main/java/com/store/reservation/store/dto/search/response/StoreDetailForManagerDto.java
@@ -27,7 +27,7 @@ public class StoreDetailForManagerDto {
     private TimeDto operationHours;
     private Location location;
     private List<LocalTime> reservationList;
-    private Long numberOfReservationPerTime;
+    private int numberOfReservationPerTime;
 
     public static StoreDetailForManagerDto from(Store store) {
         return StoreDetailForManagerDto.builder()
